added user detail fragment

+++ app/src/main/java/com/howettl/mvvm/data/model/User.kt
@@ -4,4 +4,4 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity
-data class User(@field:PrimaryKey val id: Long, val name: String, val username: String, val email: String, val phone: String, val website: String)
+data class User(@field:PrimaryKey val id: Int, val name: String, val username: String, val email: String, val phone: String, val website: String)
